nothing

+++ src/main/java/irysc/gachesefid/Controllers/ManageUserController.java
@@ -529,7 +529,7 @@ public class ManageUserController {
                     .append("phone", phone)
                     .append("status", "active")
                     .append("level", false)
-                    .append("money", config.getInteger("init_money"))
+                    .append("money", (double)config.getInteger("init_money"))
                     .append("coin", config.getDouble("init_coin"))
                     .append("student_id", Utility.getRandIntForStudentId(Utility.getToday("/").substring(0, 6).replace("/", "")))
                     .append("events", new ArrayList<>())
